RLIB-19: Provide only public API for RPDF
Move internal functions, structures and defines to rpdf-internal.h.
The unit test should only use public API to get the result buffer
and the PDF size.

Make sure it properly builds for Windows with MinGW cross compiler
under Fedora 23. Have a stable ABI for Windows by using a DEF file
to ensure fixed order of the exported symbols.
