+++ rpdf/librpdf.def
@@ -1,0 +1,34 @@
+LIBRARY "rpdf.dll"
+
+EXPORTS
+rpdf_new@1
+rpdf_set_compression@2
+rpdf_get_compression@3
+rpdf_finalize@4
+rpdf_set_title@5
+rpdf_set_subject@6
+rpdf_set_author@7
+rpdf_set_keywords@8
+rpdf_set_creator@9
+rpdf_translate@10
+rpdf_rotate@11
+rpdf_new_page@12
+rpdf_set_page@13
+rpdf_set_font@14
+rpdf_set_font_size@15
+rpdf_text_callback@16
+rpdf_text@17
+rpdf_image@18
+rpdf_link@19
+rpdf_moveto@20
+rpdf_set_line_width@21
+rpdf_lineto@22
+rpdf_closepath@23
+rpdf_stroke@24
+rpdf_rect@25
+rpdf_fill@26
+rpdf_setrgbcolor@27
+rpdf_text_width@28
+rpdf_arc@29
+rpdf_get_buffer@30
+rpdf_free@31
